chore: allow limited nix-shell for aarch64-darwin
Some packages are broken, while others can never work: udev.
Still it's still useful for simple checkins.
