+++ control-plane/agents/Cargo.toml
@@ -57,7 +57,6 @@ once_cell = "1.9.0"
 indexmap = "1.8.0"
 k8s-openapi = { version = "0.15.0", features = ["v1_20"] }
 kube = { version = "0.74.0", features = ["derive"] }
-tokio-udev = { version = "0.8.0" }
 nvmeadm = { path = "../../utils/dependencies/nvmeadm" }
 futures-util = { version = "0.3.21" }
 tokio-stream = { version = "0.1.9" }
@@ -73,6 +72,9 @@ opentelemetry = { version = "0.17.0", features = ["rt-tokio-current-thread"] }
 tracing = "0.1.35"
 tracing-subscriber = { version = "0.3.15", features = [ "env-filter" ] }
 opentelemetry-semantic-conventions = "0.9.0"
+
+[target.'cfg(target_os="linux")'.dependencies]
+tokio-udev = { version = "0.8.0" }
 
 [dev-dependencies]
 deployer-cluster = { path = "../../utils/deployer-cluster" }
