+++ control-plane/csi-driver/Cargo.toml
@@ -14,7 +14,6 @@ path = "src/bin/controller/main.rs"
 name = "csi-node"
 path = "src/bin/node/main.rs"
 
-
 [build-dependencies]
 tonic-build = "0.8.0"
 prost-build = "0.11.1"
@@ -54,20 +53,22 @@ glob = "0.3.0"
 lazy_static = "1.4.0"
 serde_json = "1.0.82"
 snafu = "0.7.1"
-sys-mount = "1.3.0"
 tower = { version = "0.4.13", features = [ "timeout", "util" ] }
-udev = "0.6.2"
 url = "2.2.2"
 uuid = { version = "0.8.2", features = ["v4"] }
 which = "4.2.2"
 k8s-openapi = { version = "0.15.0", features = ["v1_20"] }
 kube = { version = "0.74.0", features = ["runtime", "derive"] }
-devinfo = { path = "../../utils/dependencies/devinfo" }
 nvmeadm = { path = "../../utils/dependencies/nvmeadm" }
 sysfs = { path = "../../utils/dependencies/sysfs" }
 common-lib = { path = "../../common" }
 utils = { path = "../../utils/utils-lib" }
 shutdown = { path = "../../utils/shutdown" }
+
+[target.'cfg(target_os="linux")'.dependencies]
+udev = "0.6.2"
+devinfo = { path = "../../utils/dependencies/devinfo" }
+sys-mount = { version = "1.3.0", default-features = false }
 
 [dependencies.serde]
 features = ["derive"]
